+++ src/lib/components/ActionPageContainer.tsx
@@ -20,7 +20,8 @@ const ActionPageContainer = ({
     w={width}
     mx="auto"
     minH="inherit"
-    py={12}
+    pt={12}
+    pb={36}
     {...boxProps}
   >
     {children}